EDIT: move data/utils.py in util/data.py
